+++ guillotina/tests/test_attachment.py
@@ -1,14 +1,32 @@
-from guillotina.behaviors.attachment import IAttachment, IMultiAttachment
+import json
+import random
+
+import pytest
+from guillotina import task_vars
+from guillotina.behaviors.attachment import IAttachment
+from guillotina.behaviors.attachment import IMultiAttachment
 from guillotina.component import get_multi_adapter
 from guillotina.interfaces import IFileManager
 from guillotina.tests import utils
 from guillotina.transactions import transaction
 
-import json
-import random
+
+_pytest_params = [
+    pytest.param("db", marks=pytest.mark.app_settings({
+        "cloud_datamanager": "db"
+    })),
+    pytest.param("redis", marks=[
+        pytest.mark.app_settings({
+            'applications': ['guillotina.contrib.redis'],
+            "cloud_storage": "guillotina.test_package.IMemoryFileField",
+            "cloud_datamanager": "redis"
+        })
+    ])
+]
 
 
-async def test_create_content_with_behavior(container_requester):
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_create_content_with_behavior(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         response, status = await requester(
             'POST',
@@ -39,7 +57,8 @@ async def test_create_content_with_behavior(container_requester):
         assert len(response) == (1024 * 1024 * 4)
 
 
-async def test_multi_upload(container_requester):
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_multi_upload(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         response, status = await requester(
             'POST',
@@ -93,11 +112,12 @@ async def test_multi_upload(container_requester):
             obj = await container.async_get('foobar')
             behavior = IMultiAttachment(obj)
             await behavior.load()
-            assert behavior.files['key1']._blob.chunks == 2
-            assert behavior.files['key2']._blob.chunks == 2
+            assert behavior.files['key1'].chunks == 2
+            assert behavior.files['key2'].chunks == 2
 
 
-async def test_large_upload_chunks(container_requester):
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_large_upload_chunks(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         response, status = await requester(
             'POST',
@@ -134,10 +154,11 @@ async def test_large_upload_chunks(container_requester):
             obj = await container.async_get('foobar')
             behavior = IAttachment(obj)
             await behavior.load()
-            assert behavior.file._blob.chunks == 2
+            assert behavior.file.chunks == 2
 
 
-async def test_tus(container_requester):
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_tus(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         response, status = await requester(
             'POST',
@@ -202,10 +223,11 @@ async def test_tus(container_requester):
             obj = await container.async_get('foobar')
             behavior = IAttachment(obj)
             await behavior.load()
-            assert behavior.file._blob.chunks == 10
+            assert behavior.file.chunks == 10
 
 
-async def test_tus_multi(container_requester):
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_tus_multi(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         response, status = await requester(
             'POST',
@@ -270,10 +292,11 @@ async def test_tus_multi(container_requester):
             obj = await container.async_get('foobar')
             behavior = IMultiAttachment(obj)
             await behavior.load()
-            assert behavior.files['file']._blob.chunks == 10
+            assert behavior.files['file'].chunks == 10
 
 
-async def test_tus_unknown_size(container_requester):
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_tus_unknown_size(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         response, status = await requester(
             'POST',
@@ -351,10 +374,11 @@ async def test_tus_unknown_size(container_requester):
             obj = await container.async_get('foobar')
             behavior = IAttachment(obj)
             await behavior.load()
-            assert behavior.file._blob.size == offset
+            assert behavior.file.size == offset
 
 
-async def test_copy_file_ob(container_requester):
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_copy_file_ob(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         _, status = await requester(
             'POST',
@@ -376,15 +400,19 @@ async def test_copy_file_ob(container_requester):
         )
         assert status == 200
 
-        request = utils.get_mocked_request(db=requester.db)
-        with request:
-            root = await utils.get_root(request)
-            async with transaction(db=requester.db, abort_when_done=True):
-                container = await root.async_get('guillotina')
+        async with transaction(db=requester.db, abort_when_done=True) as txn:
+            root = await txn.manager.get_root(txn)
+            container = await root.async_get('guillotina')
+            task_vars.container.set(container)
+            request = utils.get_mocked_request(db=requester.db)
+            with request:
                 obj = await container.async_get('foobar')
                 attachment = IAttachment(obj)
                 await attachment.load()
-                existing_bid = attachment.file._blob.bid
+                if manager_type == 'db':
+                    existing_id = attachment.file._blob.bid
+                else:
+                    existing_id = attachment.file.uri
                 cfm = get_multi_adapter(
                     (obj, request, IAttachment['file'].bind(attachment)), IFileManager
                 )
@@ -392,10 +420,14 @@ async def test_copy_file_ob(container_requester):
                     (obj, request, IAttachment['file'].bind(attachment)), IFileManager
                 )
                 await cfm.copy(from_cfm)
-                assert existing_bid != attachment.file._blob.bid
+                if manager_type == 'db':
+                    assert existing_id != attachment.file._blob.bid
+                else:
+                    assert existing_id != attachment.file.uri
 
 
-async def test_tus_unfinished_error(container_requester):
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_tus_unfinished_error(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         _, status = await requester(
             'POST',
@@ -451,7 +483,8 @@ async def test_tus_unfinished_error(container_requester):
         assert status == 201
 
 
-async def test_tus_with_empty_file(container_requester):
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_tus_with_empty_file(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         response, status = await requester(
             'POST',
@@ -514,10 +547,11 @@ async def test_tus_with_empty_file(container_requester):
             obj = await container.async_get('foobar')
             behavior = IAttachment(obj)
             await behavior.load()
-            assert behavior.file._blob.chunks == 0
+            assert behavior.file.chunks == 0
 
 
-async def test_update_filename_on_files_on_post(container_requester):
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_update_filename_on_files_on_post(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         data = 'WFhY' * 1024 + 'WA=='
         response, status = await requester(
@@ -547,7 +581,9 @@ async def test_update_filename_on_files_on_post(container_requester):
         response, status = await requester('GET', '/db/guillotina/foobar/@download/file')
         assert len(response) == 3073
 
-async def test_update_filename_on_files(container_requester):
+
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_update_filename_on_files(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         response, status = await requester(
             'POST',
@@ -593,7 +629,8 @@ async def test_update_filename_on_files(container_requester):
         assert data['extension'] == 'jpg'
 
 
-async def test_should_fourohfour_with_invalid_fieldname(container_requester):
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_should_fourohfour_with_invalid_fieldname(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         response, status = await requester(
             'POST',
@@ -613,7 +650,8 @@ async def test_should_fourohfour_with_invalid_fieldname(container_requester):
         assert status == 404
 
 
-async def test_file_head(container_requester):
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_file_head(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         _, status = await requester(
             'POST',
@@ -642,7 +680,8 @@ async def test_file_head(container_requester):
         assert headers['Content-Type'] == 'application/octet-stream'
 
 
-async def test_file_head_not_found(container_requester):
+@pytest.mark.parametrize("manager_type", _pytest_params)
+async def test_file_head_not_found(manager_type, redis_container, container_requester):
     async with container_requester as requester:
         _, status = await requester(
             'POST',